+++ vortex-datafusion/Cargo.toml
@@ -12,7 +12,7 @@ rust-version.workspace = true
 
 [dependencies]
 vortex-array = { path = "../vortex-array" }
-vortex-dtype = { path = "../vortex-dtype", features = ["arrow"] }
+vortex-dtype = { path = "../vortex-dtype" }
 vortex-error = { path = "../vortex-error" }
 
 arrow-array = { workspace = true }
