+++ vortex-dtype/Cargo.toml
@@ -16,7 +16,6 @@ name = "vortex_dtype"
 path = "src/lib.rs"
 
 [dependencies]
-arrow-schema = { workspace = true, optional = true }
 flatbuffers = { workspace = true, optional = true }
 half = { workspace = true, features = ["num-traits"] }
 itertools = { workspace = true }
@@ -39,7 +38,6 @@ workspace = true
 [features]
 # Uncomment for improved IntelliJ support
 # default = ["flatbuffers", "proto", "serde"]
-arrow = ["dep:arrow-schema"]
 flatbuffers = ["dep:flatbuffers"]
 proto = ["dep:prost"]
 serde = ["dep:serde"]
